feat: Se creo la funcionalidad para eliminar una organizacion

+++ src/app/organizacion/lista-organizaciones/lista-organizaciones.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { OrganizacionService } from '../services/organizacion.service';
+import { trigger,style,transition,animate,keyframes,query,stagger } from '@angular/animations';
 import { ActivatedRoute } from '@angular/router';
 import { Router } from '@angular/router';
 import { PipeTransform, Pipe } from '@angular/core';
@@ -9,11 +10,31 @@ import { Organizacion } from '../../models/Organizacion';
 @Component({
   selector: 'app-lista-organizaciones',
   templateUrl: './lista-organizaciones.component.html',
-  styleUrls: ['./lista-organizaciones.component.css']
+  styleUrls: ['./lista-organizaciones.component.css'],
+  animations: [
+    trigger('animation', [
+      transition('* => *', [
+        query(':enter', style({ opacity: 0 }), {optional: true}),
+        query(':enter', stagger('200ms', [
+          animate('.6s ease-in', keyframes([
+            style({opacity: 0, transform: 'translateY(-75%)', offset: 0}),
+            style({opacity: .5, transform: 'translateY(35px)',  offset: 0.3}),
+            style({opacity: 1, transform: 'translateY(0)',     offset: 1.0}),
+          ]))]), {optional: true})
+          ,
+        query(':leave', stagger('200ms', [
+          animate('.6s ease-out', keyframes([
+            style({opacity: 1, transform: 'translateY(0)', offset: 0}),
+            style({opacity: .5, transform: 'translateY(35px)',  offset: 0.3}),
+            style({opacity: 0, transform: 'translateY(-75%)',     offset: 1.0}),
+          ]))]), {optional: true})
+      ])
+    ])
+  ]
 })
+
 export class ListaOrganizacionesComponent implements OnInit {
   organizaciones: any;
-  organizacionBuscada: String = '5a68fb901819653ad8d4ab12'; // Ejemplo
 
   constructor(
     private organizacionService: OrganizacionService,
@@ -36,18 +57,18 @@ export class ListaOrganizacionesComponent implements OnInit {
     )
   }
 
-  updateOrganizacion(): void {
-    // this.organizacionService.updateOrganizacion(this.organizacion, this.organizacion._id)
-    //   .subscribe(res => {
-        
-    //   })
-  }
-
-  deleteOrganizacion(id: String): void{
-    this.organizacionService.deleteOrganizacion(id)
-      .subscribe(res => {
-        console.log(res) 
-      })
+  deleteOrganizacion(id , i): void{
+    const response = confirm('¿Estas seguro(a) que deseas eliminarlo?');
+    if(response) {
+      this.organizacionService.deleteOrganizacion(id)
+        .subscribe(res => {
+            console.log(res),
+            this.organizaciones.organizacion.splice(i, 1);
+        },(err) => {
+          console.log(err);
+        }
+      )
+    } 
   }
 
 }
